**refactor(leaderboard): extract `extractLineValue` to helpers and clean up duplication**
- Moved `extractLineValue` logic to shared helpers for reuse.
- Removed redundant implementations in `LeaderboardTableDesktop` and `LeaderboardTableMobile`.
- Added detailed error logging in `LeaderboardDetailPage` for session storage issues.
